Move InstructionDeleter into its own header.
Add file-level comments on purpose and intended API usage.

Cleanup API comments.
